# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoroBeacon API",
    "description": "Monitoring and alerting for Stellar/Soroban contract events. All endpoints are mounted under /api/v1. The API is read-write for monitors, rules and channels; alerts are read-only.\n\nAuthentication: when API_TOKEN is configured, every endpoint below requires `Authorization: Bearer <token>` and answers 401 through the standard error envelope without one. The only exceptions are the probes /livez, /readyz and /health, which stay reachable so an orchestrator does not need the token. With API_TOKEN unset the API is open and none of this applies.",
    "version": "1.0.0",
    "license": {
      "name": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
    "/monitors": {
      "post": {
        "summary": "Create a monitor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MonitorInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Monitor" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      },
      "get": {
        "summary": "List monitors",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Case-insensitive substring match against the monitor name"
          },
          {
            "name": "enabled",
            "in": "query",
            "schema": { "type": "boolean" },
            "description": "Tri-state: omitted = all, true = enabled only, false = disabled only"
          },
          {
            "name": "sort",
            "in": "query",
            "schema": { "type": "string", "enum": ["name", "id", "created_at"], "default": "name" },
            "description": "Allowlisted order: name (default), id (newest first), created_at"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Opaque continuation cursor from a previous page"
          }
        ],
        "responses": {
          "200": {
            "description": "Monitors",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MonitorPage" }
              }
            }
          }
        }
      }
    },
    "/monitors/bulk": {
      "post": {
        "summary": "Bulk enable or disable monitors",
        "description": "Sets enabled on an explicit list of monitor IDs in one statement. Unknown IDs are returned and do not prevent the known ones from applying. There is no all-monitors shorthand.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BulkMonitorInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Applied",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BulkMonitorResult" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/monitors/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ID" }
      ],
      "get": {
        "summary": "Fetch one monitor",
        "responses": {
          "200": {
            "description": "Monitor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Monitor" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "summary": "Update a monitor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MonitorInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Monitor" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "summary": "Delete a monitor",
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/channels": {
      "post": {
        "summary": "Create a notification channel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChannelInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Channel" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      },
      "get": {
        "summary": "List channels",
        "parameters": [
          {
            "name": "enabled",
            "in": "query",
            "schema": { "type": "boolean" },
            "description": "When true, only enabled channels are returned"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Opaque continuation cursor from a previous page"
          }
        ],
        "responses": {
          "200": {
            "description": "Channels",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChannelPage" }
              }
            }
          }
        }
      }
    },
    "/channels/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ID" }
      ],
      "get": {
        "summary": "Fetch one channel",
        "responses": {
          "200": {
            "description": "Channel",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Channel" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "summary": "Update a channel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChannelInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Channel" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "summary": "Delete a channel",
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/channels/{id}/test": {
      "parameters": [
        { "$ref": "#/components/parameters/ID" }
      ],
      "post": {
        "summary": "Send a test notification through a channel",
        "responses": {
          "200": { "description": "Test sent" },
          "404": { "$ref": "#/components/responses/Error" },
          "502": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/alerts": {
      "get": {
        "summary": "List alerts",
        "parameters": [
          {
            "name": "monitor_id",
            "in": "query",
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "rule_id",
            "in": "query",
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "contract_id",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Exact match on payload.contract_id"
          },
          {
            "name": "sort",
            "in": "query",
            "schema": { "type": "string", "enum": ["created_at_desc", "created_at_asc"], "default": "created_at_desc" },
            "description": "Keyset order; anything else is 400"
          },
          {
            "name": "since",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "RFC 3339 lower bound on alert creation"
          },
          {
            "name": "until",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "RFC 3339 upper bound on alert creation"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Opaque continuation cursor from a previous page"
          }
        ],
        "responses": {
          "200": {
            "description": "Alerts",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AlertPage" }
              }
            }
          }
        }
      }
    },
    "/alerts.csv": {
      "get": {
        "summary": "Export alerts as CSV",
        "description": "Streams every alert matching the same filters as GET /alerts as text/csv. Columns are fixed: id, monitor_name, rule_id, contract_id, event_name, event_id, ledger, created_at, payload. Values beginning with =, +, - or @ are prefixed with an apostrophe to prevent spreadsheet formula injection. When no limit is given the export is capped at 10000 rows.",
        "parameters": [
          {
            "name": "monitor_id",
            "in": "query",
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "rule_id",
            "in": "query",
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "contract_id",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Exact match on payload.contract_id"
          },
          {
            "name": "from",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "RFC 3339 lower bound on alert creation"
          },
          {
            "name": "to",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "RFC 3339 upper bound on alert creation"
          },
          {
            "name": "sort",
            "in": "query",
            "schema": { "type": "string", "enum": ["created_at_desc", "created_at_asc"], "default": "created_at_desc" },
            "description": "Keyset order; anything else is 400"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "default": 10000, "maximum": 10000 },
            "description": "Maximum rows to export; the endpoint caps at 10000"
          }
        ],
        "responses": {
          "200": {
            "description": "Alerts as CSV",
            "content": {
              "text/csv": {
                "schema": { "type": "string" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/audit": {
      "get": {
        "summary": "Audit log of configuration changes",
        "description": "Append-only entries for monitor, rule and channel create, update and delete. diff records field names only, never values.",
        "parameters": [
          { "name": "target_type", "in": "query", "schema": { "type": "string", "enum": ["monitor", "rule", "channel"] } },
          { "name": "target_id", "in": "query", "schema": { "type": "integer", "format": "int64" } },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 500 } }
        ],
        "responses": {
          "200": {
            "description": "Audit entries, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["entries"],
                  "properties": {
                    "entries": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["id", "actor", "action", "target_type", "target_id", "created_at"],
                        "properties": {
                          "id": { "type": "integer", "format": "int64" },
                          "actor": { "type": "string" },
                          "action": { "type": "string", "enum": ["create", "update", "delete"] },
                          "target_type": { "type": "string", "enum": ["monitor", "rule", "channel"] },
                          "target_id": { "type": "integer", "format": "int64" },
                          "diff": { "type": "object" },
                          "created_at": { "type": "string", "format": "date-time" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/templates": {
      "post": {
        "summary": "Create a monitor template",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MonitorTemplateInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MonitorTemplate" } } }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      },
      "get": {
        "summary": "List monitor templates",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "templates": { "type": "array", "items": { "$ref": "#/components/schemas/MonitorTemplate" } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/templates/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/ID" }],
      "get": {
        "summary": "Get a monitor template",
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MonitorTemplate" } } }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "summary": "Update a monitor template",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MonitorTemplateInput" } } }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MonitorTemplate" } } }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "summary": "Delete a monitor template",
        "description": "Existing monitors created from this template are unaffected.",
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/monitors/import": {
      "post": {
        "summary": "Bulk import contracts from a file",
        "description": "Accepts JSON or CSV (text/csv). Each contract ID creates one monitor. Validates all IDs before writing. Existing contracts are skipped.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ImportInput" }
            },
            "text/csv": {
              "schema": { "type": "string", "description": "One contract ID per line" }
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Monitor name prefix (CSV mode only, defaults to 'imported')"
          }
        ],
        "responses": {
          "200": {
            "description": "Import results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": { "type": "integer" },
                    "results": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ImportRowResult" }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "One of the static tokens from API_TOKEN (comma-separated to allow rotation). A dashboard session cookie from /login is accepted in the browser's own same-origin requests."
      }
    },
    "parameters": {
      "ID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "format": "int64" }
      },
...
    },
    "responses": {
      "Error": {
        "description": "Error envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "schemas": {
...
      "Monitor": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "contract_ids": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Soroban contract strkeys (C...)"
          },
          "enabled": { "type": "boolean" },
          "channel_ids": {
            "type": "array",
            "items": { "type": "integer", "format": "int64" }
          }
        }
      },
      "MonitorInput": {
        "type": "object",
        "required": ["name", "contract_ids"],
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "contract_ids": {
            "type": "array",
            "minItems": 1,
            "items": { "type": "string", "pattern": "^C[A-Z2-7]{55}$" }
          },
          "enabled": { "type": "boolean", "default": true },
          "channel_ids": {
            "type": "array",
            "items": { "type": "integer", "format": "int64" }
          }
        }
      },
      "BulkMonitorInput": {
        "type": "object",
        "required": ["ids", "enabled"],
        "properties": {
          "ids": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": { "type": "integer", "format": "int64" }
          },
          "enabled": { "type": "boolean" }
        }
      },
...
      "MonitorPage": {
        "type": "object",
        "properties": {
          "monitors": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Monitor" }
          },
          "next_cursor": {
            "type": "string",
            "description": "Absent on the last page"
          }
        }
      },
...
    }
  }
}